Refactor for removal of chapter parsing
