+++ audiobooks.py
@@ -1,12 +1,11 @@
 """Main program for converting text files to audiobooks."""
 
 import os
-import re
 
-from text_processing import get_user_book_url, download_book_content, get_book_title, chunk_text_from_file
-from text_processing import setup_output_directory, export_raw_text, clean_text, export_cleaned_text
+from text_processing import get_user_book_url, download_book_content, get_book_title
+from text_processing import setup_output_directory, export_raw_text, clean_text
 
-# from speech_processing import create_audiobook_chapter
+from speech_processing import generate_full_audiobook
 
 def main():
     """
@@ -48,68 +47,8 @@ def main():
     # We will use this cleaned file for the text-to-speech conversion.
     cleaned_file_path = os.path.join(output_directory, f"{book_title}_cleaned.txt")
 
-    if cleaned_book_content:
-        if export_cleaned_text(cleaned_book_content, cleaned_file_path):
-            print(f"Cleaned text exported to: {cleaned_file_path}")
-
-            # # --- Chapter Parsing and Audiobook Generation Steps ---
-            # print("\n--- Starting Chapter Parsing and Audiobook Generation ---")
-
-            # # Read the cleaned text content for chapter parsing
-            # # (assuming clean_text outputs content, otherwise re-read from cleaned_file_path)
-            # # For simplicity, let's re-read the content from the cleaned file
-            # try:
-            #     with open(cleaned_file_path, "r", encoding="utf-8") as f:
-            #         full_cleaned_text = f.read()
-            # except Exception as e:
-            #     print(f"Error reading cleaned text file: {e}")
-            #     return
-
-            # chapters_data = parse_chapters(full_cleaned_text)
-
-            # if not chapters_data:
-            #     print("No chapters identified. Converting entire cleaned text as a single 'Full Book' chapter.")
-            #     chapters_data = [{
-            #         'title': f"{book_title} (Full Book)",
-            #         'content': full_cleaned_text
-            #     }]
-
-            # # Process each chapter
-            # for i, chapter in enumerate(chapters_data):
-            #     chapter_title_for_file = re.sub(r'[^\w\s-]', '', chapter['title']).replace(' ', '_')[:50] # Sanitize for filename
-            #     chapter_output_filename_base = f"{book_title}_{chapter_title_for_file}"
-
-            #     print(f"\nSynthesizing: {chapter['title']} (Chapter {i+1} of {len(chapters_data)})")
-
-            #     # Temporarily save the chapter content to a file for `create_audiobook_chapter`
-            #     # A more direct way would be to modify create_audiobook_chapter to accept content directly,
-            #     # but sticking to the file-based approach for consistency with previous discussion.
-            #     temp_chapter_filepath = os.path.join(output_directory, f"temp_chapter_{i+1}.txt")
-            #     with open(temp_chapter_filepath, "w", encoding="utf-8") as f:
-            #         f.write(chapter['content'])
-
-            #     # Call create_audiobook_chapter for each identified chapter
-            #     synthesized_audio_files = create_audiobook_chapter(
-            #         input_filepath=temp_chapter_filepath,
-            #         chapter_num=i + 1, # Use the loop index as chapter number
-            #         max_chars_per_chunk=4800 # Google Cloud TTS limit
-            #     )
-
-            #     if synthesized_audio_files:
-            #         print(f"Generated {len(synthesized_audio_files)} audio part(s) for '{chapter['title']}'.")
-            #     else:
-            #         print(f"Failed to generate audio for '{chapter['title']}'.")
-
-            #     # Clean up temporary chapter file
-            #     os.remove(temp_chapter_filepath)
-            
-            # print(f"\n--- Audiobook Generation Complete for '{book_title}' ---")
-            # print("All chapter audio files are located in the './output/' directory.")
-
-        else:
-            print("Failed to export cleaned text. Skipping audiobook generation.")
-    else:
-        print("No cleaned content was generated. Skipping export and audiobook generation.")
+    # 8. Generate full audiobook from chunked files. 
+    generate_full_audiobook(book_title, cleaned_book_content, cleaned_file_path, output_directory)
 
 if __name__ == "__main__":
     main()
